project/requirements.txt
# Streamlit and Core
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Data Science and Visualization
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# AI and NLP (Optional - for real AI integration)
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
anthropic>=0.8.0

# System and Command Execution
paramiko>=3.3.0
subprocess32>=3.5.4; python_version < '3.8'
psutil>=5.9.0

# Docker Integration
docker>=6.1.0

# Database and Storage
sqlalchemy>=2.0.0


# Networking and HTTP
requests>=2.31.0
urllib3>=2.0.0
httpx>=0.25.0

# File Handling and Utilities
pathlib2>=2.3.7; python_version < '3.4'
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.2

# Date and Time
python-dateutil>=2.8.0
pytz>=2023.3

# Logging and Monitoring
structlog>=23.2.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Performance
cachetools>=5.3.0
memory-profiler>=0.61.0

# Optional: Scientific Computing
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional: Image Processing
Pillow>=10.0.0

# Optional: Web Scraping (for advanced features)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Excel Support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional: Advanced CLI
click>=8.1.0
rich>=13.5.0

# Optional: CPU Info
py-cpuinfo>=9.0.0

# Optional: System Information
distro>=1.8.0